requirements.txt
Django==2.2.2
payu-biz==2.0
paywix==2.1
pytz==2019.1
sqlparse==0.3.0
